pyproject.toml
[project]
name = "utterance-intention-generator"
version = "0.1.0"
description = "Utterance Intention Generator using Google ADK and Gemini Flash LLM"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.2.1",
    "asyncio",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.3",
] 

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]